Reengineer DPDK DoConfig tests
- Changed class prefix from DpdkConfig to DpdkDoConfig.

- Changed file prefix from dpdk_config to dpdk_do_config.

- Stripped out all but the connect() and getPipelineConfig()
  test cases. The remaining test cases were fragile, and they
  artifically inflated the code coverage statistics by running
  code they could not verify.
