+++ ovs-p4rt/sidecar/tests/dpdk/CMakeLists.txt
@@ -76,11 +76,11 @@ macro(define_dpdk_config_test TARGET)
 endmacro(define_dpdk_config_test)
 
 #-----------------------------------------------------------------------
-# DPDK ConfigTableEntry unit tests
+# DoConfigEntry tests
 #-----------------------------------------------------------------------
 if(BUILD_CLIENT)
-  define_dpdk_config_test(dpdk_config_fdb_entry_test)
-  define_dpdk_config_test(dpdk_config_tunnel_entry_test)
+  define_dpdk_config_test(dpdk_do_config_fdb_entry_test)
+  define_dpdk_config_test(dpdk_do_config_tunnel_entry_test)
 endif()
 
 # Export list of unit tests.
