+++ ovs-p4rt/sidecar/tests/dpdk/dpdk_do_config_fdb_entry_test.cc
@@ -1,0 +1,73 @@
+// Copyright 2025 Derek Foster
+// SPDX-License-Identifier: Apache-2.0
+
+// Minimal unit test for DoConfigFdbEntry().
+
+#include <absl/status/status.h>
+
+// clang-format off
+#include <gtest/gtest.h>
+#include <gmock/gmock.h>
+// clang-format on
+
+#include "client/ovsp4rt_test_client_mock.h"
+#include "ovsp4rt/ovs-p4rt.h"
+#include "ovsp4rt_doconfig_int.h"
+#include "p4/config/v1/p4info.pb.h"
+#include "p4info_text.h"
+#include "stratum/lib/utils.h"
+
+using ::testing::Return;
+
+namespace ovsp4rt {
+
+constexpr bool INSERT_ENTRY = true;
+constexpr char GRPC_ADDR[] = "1.2.3.4:5678";
+
+class DpdkDoConfigFdbEntryTest : public ::testing::Test {
+ protected:
+  DpdkDoConfigFdbEntryTest() {}
+  ~DpdkDoConfigFdbEntryTest() = default;
+
+  void InitP4Info(::p4::config::v1::P4Info* p4info) {
+    auto status = stratum::ParseProtoFromString(P4INFO_TEXT, p4info);
+    EXPECT_TRUE(status.ok())
+        << "ParseProtoFromString: " << status;
+  }
+};
+
+/**
+ * Exercises client.connect() error path.
+ */
+TEST_F(DpdkDoConfigFdbEntryTest, connectFailure) {
+  struct mac_learning_info learn_info = {0};
+
+  TestClientMock client;
+  EXPECT_CALL(client, connect).WillOnce(Return(absl::InternalError("connect")));
+
+  auto status = DoConfigFdbEntry(client, learn_info, INSERT_ENTRY, GRPC_ADDR);
+
+  ASSERT_FALSE(status.ok());
+  ASSERT_TRUE(IsInternal(status) && status.message() == "connect")
+      << status;
+}
+
+/**
+ * Exercises config.getPipelineConfig() error path.
+ */
+TEST_F(DpdkDoConfigFdbEntryTest, getPipelineConfigFailure) {
+  struct mac_learning_info learn_info = {0};
+
+  TestClientMock client;
+  EXPECT_CALL(client, connect).WillOnce(Return(absl::OkStatus()));
+  EXPECT_CALL(client, getPipelineConfig)
+      .WillOnce(Return(absl::InternalError("getPipelineConfig")));
+
+  auto status = DoConfigFdbEntry(client, learn_info, INSERT_ENTRY, GRPC_ADDR);
+
+  ASSERT_FALSE(status.ok());
+  ASSERT_TRUE(IsInternal(status) && status.message() == "getPipelineConfig")
+      << status;
+}
+
+}  // namespace ovsp4rt
